containers: generalize Frame to Frame<I> (saved_len u32 -> I)
The semi-persistent vector's per-frame `saved_len` was stored as `u32`, capping
a frame's saved length at u32::MAX even when the vector's index type `I` is
wider (e.g. `u64`). Make `Frame` generic in the index type so `saved_len: I`
matches the vector it describes; `diff_start` stays `usize`. Drops the lossy
`as u32` / `try_from_usize(...).expect("overflow")` round-trips in mark/restore.

Also corrects the design docs + egraph A0 overview: snapshots are memory-cheap
(a sparse negative diff, not a full copy) and `mark` is O(diff), not O(1).

+++ containers/src/append_only_vec.rs
@@ -12,7 +12,7 @@ use super::token::{ContainerId, ForkHistory, VecToken};
 /// `TRACK=false` compiles out all fork history and frame tracking.
 pub struct AppendOnlyVec<T, const TRACK: bool = true> {
     data: Vec<T>,
-    frames: Vec<u32>,
+    frames: Vec<usize>,
     forks: ForkHistory,
     id: ContainerId,
 }
@@ -72,7 +72,7 @@ impl<T, const TRACK: bool> AppendOnlyVec<T, TRACK> {
             frame_index: self.frames.len() as u32,
             container_id: self.id,
         };
-        self.frames.push(self.data.len() as u32);
+        self.frames.push(self.data.len());
         token
     }
 
@@ -80,7 +80,7 @@ impl<T, const TRACK: bool> AppendOnlyVec<T, TRACK> {
         assert!(TRACK, "restore() called on untracked AppendOnlyVec");
         self.validate_token(&token);
         let target = token.frame_index as usize;
-        let saved_len = self.frames[target] as usize;
+        let saved_len = self.frames[target];
         self.data.truncate(saved_len);
         self.frames.truncate(target);
         self.forks.fork(&token, self.frames.len() as u32);
